requirements.txt
tensorflow-cpu
numpy